+++ Source/GA6thFinal_Framework/GameScripts/Scripts/Token/Token.cpp
@@ -1,12 +1,12 @@
 ﻿#include "pchScripts.h"
 #include "Token.h"
 
-UINT8 Token::GetStackCount() const
+UINT16 Token::GetStackCount() const
 {
     return _stackCount;
 }
 
-UINT8 Token::GetMaxStackCount() const
+UINT16 Token::GetMaxStackCount() const
 {
     return _maxStackCount;
 }
@@ -16,23 +16,23 @@ void Token::ClearStack()
     _stackCount = 0;
 }
 
-void Token::SetStack(UINT8 count)
+void Token::SetStack(UINT16 count)
 {
-    _stackCount = std::clamp(count, (UINT8)0, _maxStackCount);
+    _stackCount = std::clamp(count, (UINT16)0, _maxStackCount);
 }
 
-void Token::AddStack(UINT8 count)
+void Token::AddStack(UINT16 count)
 {
     _stackCount += count;
     _stackCount = std::min(_stackCount, _maxStackCount);
 }
 
-void Token::RemoveStack(UINT8 count)
+void Token::RemoveStack(UINT16 count)
 {
-    _stackCount = _stackCount <= count ? (UINT8)0 : _stackCount - count;
+    _stackCount = _stackCount <= count ? (UINT16)0 : _stackCount - count;
 }
 
-void Token::SetMaxStackCount(UINT8 maxStack)
+void Token::SetMaxStackCount(UINT16 maxStack)
 {
     _maxStackCount = maxStack;
 }
