+++ Source/GA6thFinal_Framework/GameScripts/Scripts/Token/Token.h
@@ -29,15 +29,15 @@ public:
     virtual void OnHit(CharacterBase* owner)        = 0;
 
 public:
-    UINT8   GetStackCount() const override;
-    UINT8   GetMaxStackCount() const override;
+    UINT16  GetStackCount() const override;
+    UINT16  GetMaxStackCount() const override;
     void    ClearStack();
-    void    SetStack(UINT8 count);
-    void    AddStack(UINT8 count = 1);
-    void    RemoveStack(UINT8 count = 1);
-    void    SetMaxStackCount(UINT8 maxStack);
+    void    SetStack(UINT16 count);
+    void    AddStack(UINT16 count = 1);
+    void    RemoveStack(UINT16 count = 1);
+    void    SetMaxStackCount(UINT16 maxStack);
 
 private:
-    UINT8   _stackCount = 0;
-    UINT8   _maxStackCount = UINT8_MAX;
+    UINT16  _stackCount = 0;
+    UINT16  _maxStackCount = UINT16_MAX;
 };
